+++ src/main/java/com/cucumber/utils/context/steps/CommonSteps.java
@@ -6,21 +6,10 @@ import io.cucumber.java.en.Given;
 @ScenarioScoped
 public class CommonSteps {
 
-    /**
-     * Cucumber native JSON reporting plugin does not support comments
-     * Therefore, the step from bellow simulates comments
-     *
-     * @param s
-     */
     @Given("^#[^><](.+)$")
     public void comment(String s) {
     }
 
-    /**
-     * Add collapsible support to Cucumber maven reporting plugin
-     *
-     * @param s
-     */
     @Given("^#>>(.+)$")
     public void startCollapsible(String s) {
     }