+++ src/test/java/com/cucumber/utils/context/vars/ScenarioVarsParserTest.java
@@ -1,59 +1,36 @@
 package com.cucumber.utils.context.vars;
 
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertInstanceOf;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class ScenarioVarsParserTest {
+class ScenarioVarsParserTest {
 
     private ScenarioVars scenarioVars;
 
-    @Before
-    public void init() {
+    @BeforeEach
+    void init() {
         this.scenarioVars = new ScenarioVars();
     }
 
     @Test
-    public void testScenarioVarsDoesNotAcceptInvalidNames() {
-        try {
-            scenarioVars.put("a[d", 1);
-            fail("Scenario vars does not meet restriction requirements");
-        } catch (RuntimeException e) {
-            assertTrue(e.getMessage().contains("Invalid variable name"));
-        }
-        try {
-            scenarioVars.put("a#d", 1);
-            fail("Scenario vars does not meet all restriction requirements");
-        } catch (RuntimeException e) {
-            assertTrue(e.getMessage().contains("Invalid variable name"));
-        }
-        try {
-            scenarioVars.put("a/d", 1);
-            fail("Scenario vars does not meet all restriction requirements");
-        } catch (RuntimeException e) {
-            assertTrue(e.getMessage().contains("Invalid variable name"));
-        }
-        try {
-            scenarioVars.put("a,d", 1);
-            fail("Scenario vars does not meet all restriction requirements");
-        } catch (RuntimeException e) {
-            assertTrue(e.getMessage().contains("Invalid variable name"));
-        }
-        try {
-            scenarioVars.put("a{d", 1);
-            fail("Scenario vars does not meet all restriction requirements");
-        } catch (RuntimeException e) {
+    void testScenarioVarsDoesNotAcceptInvalidNames() {
+        for (String invalid : new String[]{"a[d", "a#d", "a/d", "a,d", "a{d"}) {
+            RuntimeException e = assertThrows(RuntimeException.class, () -> scenarioVars.put(invalid, 1));
             assertTrue(e.getMessage().contains("Invalid variable name"));
         }
     }
 
     @Test
-    public void testGlobalPlaceholderFillInSimpleText() {
+    void testGlobalPlaceholderFillInSimpleText() {
         String a = "The #[animal] is running through the #[location]";
         Map<String, Object> values = new HashMap<>();
         values.put("animal", "chupacabra");
@@ -66,7 +43,7 @@ public class ScenarioVarsParserTest {
     }
 
     @Test
-    public void testGlobalPlaceholderFillInSimpleTextAndInvalidParameters() {
+    void testGlobalPlaceholderFillInSimpleTextAndInvalidParameters() {
         String a = "The #[animal] is running through the #[location]";
         Map<String, Object> values = new HashMap<>();
         values.put("animals", "chupacabra");
@@ -76,7 +53,7 @@ public class ScenarioVarsParserTest {
     }
 
     @Test
-    public void testScenarioPlaceholderFillInSimpleText() {
+    void testScenarioPlaceholderFillInSimpleText() {
         String a = "The #[animal] is running through the #[location]";
         Map<String, Object> values = new HashMap<>();
         values.put("animal", "chupacabra");
@@ -86,7 +63,7 @@ public class ScenarioVarsParserTest {
     }
 
     @Test
-    public void testScenarioPlaceholderFillInSimpleTextAndInvalidParameters() {
+    void testScenarioPlaceholderFillInSimpleTextAndInvalidParameters() {
         String a = "The #[animal] is running through the #[location]";
         Map<String, Object> values = new HashMap<>();
         values.put("animals", "chupacabra");
@@ -96,7 +73,7 @@ public class ScenarioVarsParserTest {
     }
 
     @Test
-    public void testGlobalAndScenarioPlaceholderFillInSimpleText() {
+    void testGlobalAndScenarioPlaceholderFillInSimpleText() {
         String a = "The #[animal] is running through the #[location]";
         Map<String, Object> values = new HashMap<>();
         values.put("animal", "chupacabra");
@@ -106,66 +83,116 @@ public class ScenarioVarsParserTest {
     }
 
     @Test
-    public void testGlobalAndScenarioPlaceholderFillInJson() {
+    void testGlobalAndScenarioPlaceholderFillInJson() {
         Map<String, Object> values = new HashMap<>();
         values.put("val1", "result 1");
         values.put("val2", "result 2");
         values.put("val3", "result 3");
         values.put("val4", "result 4");
         values.put("location", "forest");
-        String actual = "[\n" + "  {\n" + "    \"_id\": \"5b48fcdd5c3da373f5114ec7\",\n"
-                + "    \"index\": 0,\n"
-                + "    \"guid\": \"ed43e5b1-3da2-4ebe-baa5-e44d4eb18a5c\",\n"
-                + "    \"isActive\": true,\n" + "    \"balance\": \"$3,812.47\",\n"
-                + "    \"picture\": \"http://placehold.it/32x32\",\n" + "    \"age\": 25,\n"
-                + "    \"eyeColor\": \"#[val3]\",\n" + "    \"name\": \"Tonya Schneider\",\n"
-                + "    \"gender\": \"female\",\n" + "    \"company\": \"PHARMEX\",\n"
-                + "    \"email\": \"tonyaschneider@pharmex.com\",\n"
-                + "    \"#[val1]\": \"#[val2]\",\n"
-                + "    \"address\": \"195 Hubbard Place, #[val1], New Hampshire, 1382\",\n"
-                + "    \"about\": \"Occaecat laboris eu #[val1] fugiat. In dolore dolore esse voluptate. Amet ipsum id nisi nulla pariatur do dolore dolore aliquip qui laboris. Aute consequat tempor incididunt sunt voluptate laboris. Velit adipisicing nostrud laboris labore eiusmod. Dolore sint laborum culpa nulla eu sunt excepteur.\\r\\n\",\n"
-                + "    \"registered\": \"2015-05-21T05:11:55 -03:00\",\n"
-                + "    \"latitude\": 67.096081,\n" + "    \"longitude\": 36.71768,\n"
-                + "    \"tags\": [\n" + "      \"aute\",\n" + "      \"dolore\",\n"
-                + "      \"#[val2]\",\n" + "      \"officia\",\n" + "      \"enim\",\n"
-                + "      \"aliqua\",\n" + "      \"~[var3]\"\n" + "    ],\n"
-                + "    \"friends\": [\n" + "      {\n" + "        \"id\": 0,\n"
-                + "        \"name\": \"~[val1] Hart\"\n" + "      },\n" + "      {\n"
-                + "        \"id\": 1,\n" + "        \"name\": \"Velazquez Sargent\"\n"
-                + "      },\n" + "      {\n" + "        \"id\": 2,\n"
-                + "        \"name\": \"De~[var4]sa\"\n" + "      }\n" + "    ],\n"
-                + "    \"greeting\": \"Hello, Tonya Schneider! You have 9 unread messages.\",\n"
-                + "    \"~[val1]\": \"banana\"\n" + "  }\n" + "]";
-        String expected = "[\n" + "  {\n" + "    \"_id\": \"5b48fcdd5c3da373f5114ec7\",\n"
-                + "    \"index\": 0,\n"
-                + "    \"guid\": \"ed43e5b1-3da2-4ebe-baa5-e44d4eb18a5c\",\n"
-                + "    \"isActive\": true,\n" + "    \"balance\": \"$3,812.47\",\n"
-                + "    \"picture\": \"http://placehold.it/32x32\",\n" + "    \"age\": 25,\n"
-                + "    \"eyeColor\": \"result 3\",\n" + "    \"name\": \"Tonya Schneider\",\n"
-                + "    \"gender\": \"female\",\n" + "    \"company\": \"PHARMEX\",\n"
-                + "    \"email\": \"tonyaschneider@pharmex.com\",\n"
-                + "    \"result 1\": \"result 2\",\n"
-                + "    \"address\": \"195 Hubbard Place, result 1, New Hampshire, 1382\",\n"
-                + "    \"about\": \"Occaecat laboris eu result 1 fugiat. In dolore dolore esse voluptate. Amet ipsum id nisi nulla pariatur do dolore dolore aliquip qui laboris. Aute consequat tempor incididunt sunt voluptate laboris. Velit adipisicing nostrud laboris labore eiusmod. Dolore sint laborum culpa nulla eu sunt excepteur.\\r\\n\",\n"
-                + "    \"registered\": \"2015-05-21T05:11:55 -03:00\",\n"
-                + "    \"latitude\": 67.096081,\n" + "    \"longitude\": 36.71768,\n"
-                + "    \"tags\": [\n" + "      \"aute\",\n" + "      \"dolore\",\n"
-                + "      \"result 2\",\n" + "      \"officia\",\n" + "      \"enim\",\n"
-                + "      \"aliqua\",\n" + "      \"~[var3]\"\n" + "    ],\n"
-                + "    \"friends\": [\n" + "      {\n" + "        \"id\": 0,\n"
-                + "        \"name\": \"~[val1] Hart\"\n" + "      },\n" + "      {\n"
-                + "        \"id\": 1,\n" + "        \"name\": \"Velazquez Sargent\"\n"
-                + "      },\n" + "      {\n" + "        \"id\": 2,\n"
-                + "        \"name\": \"De~[var4]sa\"\n" + "      }\n" + "    ],\n"
-                + "    \"greeting\": \"Hello, Tonya Schneider! You have 9 unread messages.\",\n"
-                + "    \"~[val1]\": \"banana\"\n" + "  }\n" + "]";
+        String actual = """
+                [
+                  {
+                    "_id": "5b48fcdd5c3da373f5114ec7",
+                    "index": 0,
+                    "guid": "ed43e5b1-3da2-4ebe-baa5-e44d4eb18a5c",
+                    "isActive": true,
+                    "balance": "$3,812.47",
+                    "picture": "http://placehold.it/32x32",
+                    "age": 25,
+                    "eyeColor": "#[val3]",
+                    "name": "Tonya Schneider",
+                    "gender": "female",
+                    "company": "PHARMEX",
+                    "email": "tonyaschneider@pharmex.com",
+                    "#[val1]": "#[val2]",
+                    "address": "195 Hubbard Place, #[val1], New Hampshire, 1382",
+                    "about": "Occaecat laboris eu #[val1] fugiat. In dolore dolore esse voluptate. Amet ipsum id nisi nulla pariatur do dolore dolore aliquip qui laboris. Aute consequat tempor incididunt sunt voluptate laboris. Velit adipisicing nostrud laboris labore eiusmod. Dolore sint laborum culpa nulla eu sunt excepteur.\\r\\n",
+                    "registered": "2015-05-21T05:11:55 -03:00",
+                    "latitude": 67.096081,
+                    "longitude": 36.71768,
+                    "tags": [
+                      "aute",
+                      "dolore",
+                      "#[val2]",
+                      "officia",
+                      "enim",
+                      "aliqua",
+                      "~[var3]"
+                    ],
+                    "friends": [
+                      {
+                        "id": 0,
+                        "name": "~[val1] Hart"
+                      },
+                      {
+                        "id": 1,
+                        "name": "Velazquez Sargent"
+                      },
+                      {
+                        "id": 2,
+                        "name": "De~[var4]sa"
+                      }
+                    ],
+                    "greeting": "Hello, Tonya Schneider! You have 9 unread messages.",
+                    "~[val1]": "banana"
+                  }
+                ]""";
+        String expected = """
+                [
+                  {
+                    "_id": "5b48fcdd5c3da373f5114ec7",
+                    "index": 0,
+                    "guid": "ed43e5b1-3da2-4ebe-baa5-e44d4eb18a5c",
+                    "isActive": true,
+                    "balance": "$3,812.47",
+                    "picture": "http://placehold.it/32x32",
+                    "age": 25,
+                    "eyeColor": "result 3",
+                    "name": "Tonya Schneider",
+                    "gender": "female",
+                    "company": "PHARMEX",
+                    "email": "tonyaschneider@pharmex.com",
+                    "result 1": "result 2",
+                    "address": "195 Hubbard Place, result 1, New Hampshire, 1382",
+                    "about": "Occaecat laboris eu result 1 fugiat. In dolore dolore esse voluptate. Amet ipsum id nisi nulla pariatur do dolore dolore aliquip qui laboris. Aute consequat tempor incididunt sunt voluptate laboris. Velit adipisicing nostrud laboris labore eiusmod. Dolore sint laborum culpa nulla eu sunt excepteur.\\r\\n",
+                    "registered": "2015-05-21T05:11:55 -03:00",
+                    "latitude": 67.096081,
+                    "longitude": 36.71768,
+                    "tags": [
+                      "aute",
+                      "dolore",
+                      "result 2",
+                      "officia",
+                      "enim",
+                      "aliqua",
+                      "~[var3]"
+                    ],
+                    "friends": [
+                      {
+                        "id": 0,
+                        "name": "~[val1] Hart"
+                      },
+                      {
+                        "id": 1,
+                        "name": "Velazquez Sargent"
+                      },
+                      {
+                        "id": 2,
+                        "name": "De~[var4]sa"
+                      }
+                    ],
+                    "greeting": "Hello, Tonya Schneider! You have 9 unread messages.",
+                    "~[val1]": "banana"
+                  }
+                ]""";
         scenarioVars.putAll(values);
         String result = ScenarioVarsParser.parse(actual, scenarioVars).toString();
-        assertEquals(result, expected, result);
+        assertEquals(expected, result);
     }
 
     @Test
-    public void testSimpleSpel() {
+    void testSimpleSpel() {
         String s = "#{T(java.net.IDN).toASCII('testá.com')}";
         assertEquals("xn--test-8na.com", ScenarioVarsParser.parse(s, scenarioVars));
         s = "Domain #{T(java.net.IDN).toASCII('testá.com')}-> idn";
@@ -173,7 +200,7 @@ public class ScenarioVarsParserTest {
     }
 
     @Test
-    public void testInvalidSpel() {
+    void testInvalidSpel() {
         String s = "This is: #{T(java.net.I).toASCII('testá.com')}";
         assertEquals("This is: #{T(java.net.I).toASCII('testá.com')}", ScenarioVarsParser.parse(s, scenarioVars));
 
@@ -185,13 +212,13 @@ public class ScenarioVarsParserTest {
     }
 
     @Test
-    public void testSpelGeneratesObject() {
+    void testSpelGeneratesObject() {
         String s = "#{new Long(1000)}";
-        assertTrue(ScenarioVarsParser.parse(s, scenarioVars) instanceof Long);
+        assertInstanceOf(Long.class, ScenarioVarsParser.parse(s, scenarioVars));
     }
 
     @Test
-    public void testSpelWithScenarioVars() {
+    void testSpelWithScenarioVars() {
         scenarioVars.put("a", 1000);
         scenarioVars.put("b", 1);
         String s = "#{#[a]+#[b]}";
@@ -199,13 +226,13 @@ public class ScenarioVarsParserTest {
     }
 
     @Test
-    public void testEmptySpel() {
+    void testEmptySpel() {
         String s = "#{}";
         assertEquals("#{}", ScenarioVarsParser.parse(s, scenarioVars));
     }
 
     @Test
-    public void testValidSpelExpWithEmbeddedInvalidSpelExp() {
+    void testValidSpelExpWithEmbeddedInvalidSpelExp() {
         String s = "#{'a#{bc'+'def'}";
         assertEquals("a#{bcdef", ScenarioVarsParser.parse(s, scenarioVars));
         String s1 = "#{'a#{b}c'+'def'";
@@ -213,7 +240,7 @@ public class ScenarioVarsParserTest {
     }
 
     @Test
-    public void testSpelWithEscapedBraces() {
+    void testSpelWithEscapedBraces() {
         String s = "#{'a\\}bc'+'d\\}ef'}";
         assertEquals("a\\}bcd\\}ef", ScenarioVarsParser.parse(s, scenarioVars));
         s = "#{'a\\#{bc'+'d\\}ef'}";
@@ -221,7 +248,7 @@ public class ScenarioVarsParserTest {
     }
 
     @Test
-    public void testSpelWithScenarioVarsHavingBraces() {
+    void testSpelWithScenarioVarsHavingBraces() {
         scenarioVars.put("myJson", "{\"a\":1}");
         String s = "#{T(io.json.compare.util.JsonUtils).toJson('#[myJson]').get('a').asInt()}";
         assertEquals(1, ScenarioVarsParser.parse(s, scenarioVars));