+++ src/test/java/com/cucumber/utils/features/stepdefs/match/NegativeMatchSteps.java
@@ -5,7 +5,7 @@ import com.google.inject.Inject;
 import io.cucumber.guice.ScenarioScoped;
 import io.cucumber.java.en.Given;
 
-import static org.junit.Assert.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @ScenarioScoped
 public class NegativeMatchSteps {