+++ src/test/java/com/cucumber/utils/features/stepdefs/paramtype/ParamTypeSteps.java
@@ -5,8 +5,8 @@ import com.google.inject.Inject;
 import io.cucumber.guice.ScenarioScoped;
 import io.cucumber.java.en.When;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 @ScenarioScoped
 public class ParamTypeSteps {