+++ src/test/java/com/cucumber/utils/features/stepdefs/placeholders/PlaceholderParserSteps.java
@@ -6,7 +6,7 @@ import io.cucumber.guice.ScenarioScoped;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @ScenarioScoped
 public class PlaceholderParserSteps {